requirements.txt
python-telegram-bot
groq
pytz